raith data now works

+++ include/gdstk/raithdata.hpp
@@ -9,13 +9,13 @@
 namespace gdstk {
 
 struct RaithData {
-    uint64_t dwelltime_selection;
+    uint8_t dwelltime_selection;
     double pitch_parallel_to_path;
     double pitch_perpendicular_to_path;
     double pitch_scale;
-    uint64_t periods;
-    uint64_t grating_type;
-    uint64_t dots_per_cycle;
+    int32_t periods;
+    int32_t grating_type;
+    int32_t dots_per_cycle;
 
     void* owner;
